GameServer Controller that only defined the CRD on startup
This implementation just builds the basic foundation of the controller
for this Custom Resource Definition. It will check and see if the
CRD has been defined, and it not, create it in the Kubernetes cluster
it resides in, and wait until it reached an Established state.

Then it just blocks until it gets told to terminate via a process signal.