Go SDK for integration with Game Servers.
First implementation was done in Go as prototyping.
Implementation is done with gRPC so it should be
relatively trivial to generate new SDKs for each
language.

Implemented gRPC sidecars server to accompany the SDK
and processing of the change in states, such that the
IP and Port of the GameServer is available in the
GameServer status when its status is Ready.

Closes #16