C++ SDK implementation, example and doc
This includes a very basic "server" that
integrates with the SDK to show how it works
as well as expansion of the SDK docs to explain
how they work.

This also includes working the SDK code gen
and compilation (on linux at least) into the
main build process as well.

Closes #7