Finalizer for GameServer until backing Pods are Terminated
Because we allocate a port to a GameSever, we cannot
delete the GameServer until that port is free again.

This means that a GameServer must not be hard deleted until
the backing Pod is successfully Terminated - and the port is
now free again.

This implements a finalizer and backing controller code to manage
this.

This is work that need to occur for #14 (Dynamic Port Allocation)
to be completed.