Development and Deployment to Minikube
You can now develop Agon, and deploy Agon to minikube
for local testing and development.

The major issue was that Minikube doesn't expose the
ExternalIP for the Node it runs - so Agon will now
fallback to the InternalIP (with a warning) if it can't
find the ExternalIP.

*Breaking change* - Agon build tools will now look in
~/.kube (like all other kubernetes tools) for Kubernetes
credentials. This was to streamline auth of minikube and other
Kubernetes clusters by conforming to what Kubernetes tooling
usually expects. If you are running a GKE cluster, you will
need to run `make gcloud-auth-cluster` again to re-download
the kubectl credentials.

Closes #30