Refactor to centralise controller
_Breaking Changes_

We are going to end up with multiple controllers
and CRDs in this system, so refactoring gameserver
code under `pkg`, and moving binaries under `cmd`.

This will keep the top level directories clean and
understandable no matter how many extra CRDs we add.

The `gameserver-controller` image is also  now the
`agon-controller` image.

Upgrade should be a delete the old
`gameserver-controller` deployment and install
as per normal.