Integrated Helm into the `build` and development system
This removes the `sed` based templating and replaces it with a
helm based workflow. All the `make` commands have stayed exactly
the same.

This also moves the existing 0.1 release helm chat to the 0.2
development chart, and includes the features that were otherwise
missing for 0.2

If you have an existing development cluster, it may be easiest
to delete them and recreate them, as the tooling will install
helm automatically now on cluster creation.

I also snuck in some small Kubernetes 1.9.x upgrades as well.

Parent ticket: #101