Generate the install.yaml from `helm template`
This implements a `make gen-install` target that will generate the
install.yaml via `helm template`. This also extends `make test`
to have a test to ensure that the `install.yaml` is always kept
up to sync with the Helm chart, otehrwise the build will fail.

This does remove the explicit permissions for RBAC, but we don't
have a specific need for them right now, so I feel it's worth the
sacrifice to ensure these two configuration options stay in sync.

Blocks on #170, Parent ticket: #101