Archive C++ src on build and release
Previously we only kept the C++ build artifacts for each build and
release, which (if my understanding of C++ is correct - which I don't
think it was a while ago) is relatively useless, unless you are on the
same OS, compiler etc.

So here we are taking a zip of the src, and keeping it, pushing it to
Cloud Storage when a PR is made, and also making sure it's in the
releases folder on release.