Upgrade gRPC to 1.12.0
This also replaces the C++ build image we have from the gRPC docker hub,
with one that we manage ourselves, since getting updates in a timely
matter was getting difficult.

The C++ base will likely go away once the transition to REST based SDK,
so the base may end up going away.

This is prerequisite for #240