Expose SDK-Server at HTTP+JSON
Implement grpc-gateway in front of the gRPC based sdk-server, so that
it can be access via HTTP+JSON.

This includes documentation and a swagger/openapi specificiation.

This also has been implemented such that the sdk-server is still a single
binary, and as such, the HTTP+JSON interface can still be used for local
development.

Closes #240