Allocation is broken when using the generated go client
When attempting to create a fleet allocation through the generated go client
the following error would occur:

`"error: "Internal error occurred: Internal error occurred: jsonpatch replace operation does not apply: doc is missing key: /status/GameServer"`

This is because we didn't mark pointers in `FleetAllocation` to be json,
`omitempty`.

When not using the client (either yaml, or rest) the following JSON is sent:

```json
{
    "apiVersion": "stable.agones.dev/v1alpha1",
    "kind": "FleetAllocation",
    "metadata": {
        "generateName": "simple-udp-",
        "namespace": "default"
    },
    "spec": {
        "fleetName": "simple-udp"
    }
}
```

Previous to this fix, when calling through the go client, the following JSON
would get generated:

```json
{
    "apiVersion": "stable.agones.dev/v1alpha1",
    "kind": "FleetAllocation",
    "metadata": {
        "creationTimestamp": null,
        "generateName": "allocatioon-",
        "namespace": "default"
    },
    "spec": {
        "fleetName": "simple-fleet-6fb7c",
        "metadata": {}
    },
    "status": {
        "GameServer": null
    }
}
```

That `nil` `GameServer` value messes up the library that creates the JSONPatch.

Now we have the `omitempty` declarations in the correct places, this is
working correctly. We also now have e2e tests to make sure the issue
does not end up replicated.

Need some user testing to determine if a hotfix is appropriate for this bug,
or if a workaround can be applied/this library can be used without a need
for a complete redeployment.