Mark GameServer Unhealthy if allocated HostPort isn't available
This marks a GameServer as unhealthy is the allocated port isn't available
on a given Node for whatever reason.

Since it's marked as Unhealthy, the Fleet controller will delete it, and the new
GameServer will be allocated a new random port that it deems open and available.

I've yet to see this in the wild,  but I had mild concerns, so figured I'd fix it
now rather than later.