Ports should always be allocated to a GameServer
The allocator was written before we have the autoscaler, which has a
requirement to be able to start GameServers past the current number of nodes,
so that the autoscaler can find room for them automaticaly.

The previous port allocator would only allow port ranges for each current node
- but this PR will allow it to go past that, and dynamically add port ranges as
required.

This also removes the requirement to do a full `GameServer` port sync every time
a Node gets removed, and drops a somewhat lengthy locking process.