Add .gocache directory for WSL users
Testing Agones on Windows using the Linux subsystem would fail with a mount denied error for the .gocache directory. This error was misleading as it was due to the directory not existing. Adding in the step of making the directory fixes the problem, and does not harm the functionality of building and testing on a native linux machine.