Lightweight library for implementing APIServer extensions
The other API extension libraries were too opinionated,
or not flexible enough to provide exactly what we need
-- register a single verb in the API, not have storage, give full control over
the http lifecycle.

This provides just enough functionality for what we need, while also providing
extension points in case we need to expand the API surface area in the future -
such as a more explict OpenAPI spec.

This is the foundational work that is required for converting
GameServerAllocations into an API Extension - next step will
be to use this library to implement that functionality.