Simplify homepage messaging
Pulling back on some of the Kubernetes terminology, and using some more
generic language, to try and reach a slightly wider audience who (a)
don't necessarily know Kubernetes very well, or possibly also don't know
(b) dedicated game servers are used for multiplayer games either.

Also trying to keep things to shorter sentences, that are easier to parse
and understand.

Also included a "Companies that use Agones" section, to setup for
external parties to submit PRs to add to the list.