Convert C++ Example to Docker Build Pattern
While this does make it a longer build time, it makes it easier to get
started with the C++ example because there is no tooling dependencies
other than Docker (Make is even optional, it just sets the default
arguments up).

Also provides an end to end example of compiling the SDK, installing it
and then using it with another C++ project.

Co-authored-by: Dmitry Sazonov <dmitry.sazonov@globant.com>