Flaky: Fix test for TestGameServerUnhealthyAfterDeletingPod
Change in behaviour on Unhealthy GameServers that are not in Fleets
(deleted vs just being marked as Unhealthy), caused this e2e test
to be flaky.

Added a TODO to review if we are still happy with this, since
release is coming up, and having stable tests is most important right
now.