Prevent race conditions by syncing cache on new Allocation elements
Added cache.InformerSynced members for the new resources that are
worked on within this code for the multicluster allocations, to ensure
they aren't empty values on controller startup.

Also applying these at the test level, so we don't get random test
flakiness due to cache's not being populated in time for the test to
run.