PortPolicy of Passthrough - Same Port for Container and Host
This implements a new PortPolicy, such that the containerPort that is
specified for the game server is dynamically set to the same random
value that the hostPort is.

This is useful for game servers that have already been configured to
broadcast the port that they originally started on as their connection
port, with no overrides.

Closes #294