Flaky: TestGameServerPassthroughPort
TestGameServerPassthroughPort exposed a pretty nasty race condition in
the sdk sidecar server, in which a request for GameServer information
could come in just before it had cached its details, resulting in a
"Not found" response.

This uses a WaitGroup to requests for GameServer information
until such time as it is initially loaded.