Flaky: Allocator bad tls certificate
This should fix the bad tls e2e test failures, as it will force
a reload of the allocator deployment - which will load the most
up to date generated ssl certificate.

This is the same implementation that is in place for the controller
as well for its webhook ssl certificates.