Add allocations metrics (#963)
This adds allocation latency metric and instruments the allocator service with OpenCensus. I've also updated the allocator deployment with Kubernetes health checking, metrics and scraping. Grafana dashboard are updated and a new allocation one is available.

This additionally fixes some issues with the last version of Kind within the build Makefile.