GameServer container restart before Ready, move to Unhealthy state After
This brings our implementation inline with what our health checking
documentation states that we do.

To solve this, we store the currently running GameServer containerID
(which is unique to that running instance), as an annotation on the
GameServer Pod.

When the annotation is not there, we know the Pod is not yet Ready,
so we can ignore it when our Unhealthy check occurs, and restarts
of the GameServer container can happen as per usual.

When the annotation is there, we know to check for failure, but to
avoid ContainerStatus.LastTerminationState polluting the result
since that crash/failure may have happened before the GameServer is
ready, we can compare the current ContainerID to the one stored in the
annotation -- which when equal means we can skip looking at the
LastTerminationState, as it was before the GameServer was marked as
Ready.

Lots of unit and e2e test updates to go with this to test it out.

Closes #956