Changed remote cluster to validate PEM certificate instead of DER (#1066)
Currently allocator service is expecting TLS certificate of PEM format and the client is expecting server TLS signing certificate of DER format. This discrepancy was not detected because e2e tests are skipping cert validation and we don't have a good test coverage for multi-cluster allocation scenarios.

This only impact the multi-cluster allocation feature which is in v1alpha1. The issue was detected while writing e2e tests for multi-cluster allocation.