Improve fleet controller response times in busy clusters.
In a cluster with significant number of fleets (> workers in Fleet
controller) and many configuration updates in flight,
it is possible for change notifications to be queued with
ever-increasing exponential back-off (up to 1000 seconds).

Because fleet controller listens for both user-triggered changes and
GSS changes, it can get really busy when allocations are in flight,
leading to slower and slower fleet reaction times to both scaling and
reflecting GSS status.

When this happens and a user script scales a fleet,
it can take multiple minutes for Fleet controller to respond.

This change caps the queueing time for each fleet at 3 seconds to avoid
spinning CPU cycles for really busy GSSs, but still give reasonable
latency for user-triggered changes.

Fixes #1107