Flaky: TestGameServerRestartBeforeReadyCrash
There was a race condition in which if a game server container crashes,
and then restarts and moves to ready, the crashed game server
container id could be the recorded container that was added as an
Annotation to the GameServer, which would is incorrect.

This fix has a check in place to ensure the game server container id
that is being added as the Annotation value, is from an actual running
container.

Closes #1173