Revert: Make it possible to create a Fleet with 0 replicas
This reverts some of the changes in #1168, which introduced a breaking
change that defaulted Fleets to have 1 replica.

Due to the nature of Go, this also meant that if you attempted to
create a Fleet with 0 replicas, it would change it to being 1 replica.

Implemented this fix to ensure there is no breaking change in the release
and its also useful to be able to create Fleets with 0 replicas in them.

Also including an extra e2e test to make sure creating a Fleet with 0
replicas continues to function into the future.