Stubs for SDK alpha/beta/stable functionality
This gets the requisite pieces in place, such that alpha and beta
features can be written for the GameServer SDKs.

This includes:
- Moving the sdk.proto into the proto directory
- Updating all the gen scripts for clients for the new sdk.proto
  location
- creating new alpha.proto and beta.proto definitions, with an empty
  service stub
- Integrated the new alpha and beta service stubs into the sdkserver
  local and sidecar implementations.
- Moved all the swagger generated files under ./sdks/swagger (felt like
  an appropriate place)

I've deliberately not update all the clients as (a) I felt it would make
this PR too big and (b) each language's client may have slightly
different integration requirements. I figured it could be handled on a
case by case basis, as required.