Release 1.3.0-rc
This includes some update to the release checklist for the new npm
publishing tools.

There is a reference to a target that has yet to be written
(`publish-sdk-packages`), but will be done with the upcoming release
right after this is merged, and will come in a following PR. Basically,
to test it I need to run it at this stage, so it makes sense to do it
together.