Super Tux Kart Example
Had to do some interesting wrangling to get this to work, but a
combination of looking at stdout and the generated logfile provides
enough information to be able to know when the server is up and running.

Would be good to eventually take this one step further and setup some AI
bots to connect as well.