Workerqueue IsConflixt needed to check error Cause
Because errors are often wrapped, when checking Kubernetes error types
on anything other than errors comming straight from Kubernetes, we
should be using `errors.Cause(err)` to find the root error.

Otherwise, this check continues to passes down to the error handler,
which defeated the purposes of the original fix.