Alpha SDK.SetPlayerCapacity & GetPlayerCapacity
This PR doesn't complete this functionality, but was enough to warrant
review.

It does two things:
1. The refactoring to move alpha and beta SDK Server implementations
into the same file as the main SDK Server, as being able to reuse the
current logic made sense on review.
2. Initial Implementation of SDK GetPlayerCapacity & SetPlayerCapacity
methods with unit tests. E2E tests can be implemented once #1397 is
resolved.

Next steps will be to implement the local SDK functionality, and
conformance tests for Go for these two functions.

Work on #1033