Flaky: TestGameServerReserve
The idea behind this solution is that as you get down into the WaitFor()
commands, there is a goroutine that is run to implement polling.

Normally, if there is a delay due to goroutine scheduling, this is fine,
as we are waiting for something to happen, and if the test comes in
after the thing happened - then that is fine.

In this instance, we definitely need to check right away, so I
implemented a special case of using a for loop instead, so that the test
can't be non-determinately delayed.