Alpha GameServer attributes added to SDK
Rather than recreating SDK.GameServer() and SDK.WatchGameServer() in the
alpha SDK, I thought it would be easiest both in the short and long
term to mimic what we have done in the GameServer CRD, and populate the
Alpha attributes from these SDK methods when the feature flag for
Player Tracking is enabled.

This is done by adding logic to the conversion function that converts
the K8s GameServer definition into the SDK GameServer definition.

This includes regenerated gRPC code for the supported clients.

I also had to fix the line endings in the C# gen.sh, so including that
in this PR.