Update Terraform to Helm 3
Upgrading the Terraform configurations to Helm 3 - mostly by creating a
new Helm 3 module, so we still have backward compatibility.

Also the new Helm 3 model still has the same variables as the Helm 2
module, although there are some changes to defaults to align more with
production workloads.

Updates to documentation were minimal, as not a huge amount of things
changed from a Terraform perspective.

Work on #1436