Flaky: TestSDKServerWatchGameServerFeatureSDKWatchSendOnExecute
Switching default: case to a timeout, since due to the asynchronous
nature of event processing, it could take a moment or two for the extra
event to be processed.

This is particularly true on the CI system, where there is a lot of
processing going on, and it may take more time than usual to process.