Update Hugo and Docsy
Due to #1668, saw that Hugo can fall over with:
`fatal error: concurrent map read and map write`

When there are lots of CPUs, so decided it was time to take the
opportunity to update Hugo.

Only big difference was to enable "unsafe" in the markdown engine, so
that it would allow HTML in markdown.