Build Terraform: Use docker image project default
This fixes the issue if your local gcloud setup points to a different
project than the build shell, but pulling the current project from the
build image configuration.

Also snuck in the change that we don't need to specifically remove the
helm module before removing the entire cluster as well.